remodeled for and implemented pip packaging
