+++ Dockerfile
@@ -5,8 +5,8 @@ COPY requirements.txt .
 RUN pip3 install -r requirements.txt
 
 # Copy files (copy api.py to main.py as the default api launcher):
-COPY ./api.py /app/main.py
-COPY ./modules /app/modules
-COPY ./data_files /app/data_files
+COPY ./dmx/api.py /app/main.py
+COPY ./dmx/modules /app/modules
+COPY ./dmx/data_files /app/data_files
 
 WORKDIR "/app"
