+++ MANIFEST.in
@@ -1,0 +1,6 @@
+include LICENSE
+include README.rst
+
+recursive-include tests *
+recursive-exclude * __pycache__
+recursive-exclude * *.py[co]
