+++ dmx/__init__.py
@@ -1,0 +1,5 @@
+"""Top-level package for dmx."""
+
+__author__ = """Robnet Kerns"""
+__email__ = "robnet@jataware.com"
+__version__ = "0.0.1"
