--- parameters.json
@@ -1,9 +1,0 @@
-{
-	"repositories": [
-		{ "path": "C:\\Users\\ehsadmin\\source\\repos\\dummy-model" },
-		{ "path": "C:\\Users\\ehsadmin\\source\\repos\\FabFlee" },
-		{ "path": "C:\\Users\\ehsadmin\\source\\repos\\FSC-WorldModelers" },
-		{ "path": "C:\\Users\\ehsadmin\\source\\repos\\MaxHop" },
-		{ "path": "C:\\Users\\ehsadmin\\source\\repos\\pythia" }
-	]
-}
